fix window closing behaviour
We reference the focused window and current workspace instead of referencing the information from the window.
This is because the window return incomplete information, leading to broken behavior.

This is definitely not perfect and will probably not be an issue when implementing this behavior in the actual shell,
but for the extension, it's "good enough."